Attempt to fix win32 errors

+++ pool.cpp
@@ -1,5 +1,6 @@
 #include "pool.h"
 #include <math.h>
+#include <stdio.h>
 
 static std::vector<Pool> pools;
 
@@ -34,9 +35,9 @@ PoolView get_pool(size_t min_size)
 
     // No pools large enough. Create a new one with size of the next
     // power of two.
-    float fp2 = log2(min_size);
-    size_t target_size = 1 << (int)ceil(fp2);
-    printf("Pools exhausted! Creating pool of size %lu to meet request of size %lu.\n", target_size, min_size);
+    float fp2 = log2f((float)min_size);
+    int target_size = 1 << (int)ceil(fp2);
+    printf("Pools exhausted! Creating pool of size %d to meet request of size %zu.\n", target_size, min_size);
     add_pool_of_size(target_size);
     return PoolView(pools[pools.size() - 1]);
 }
